Refactor tool_shed.util.tool_dependency_util into galaxy.
